requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
vanna 
httpx
vanna[chromadb,ollama,postgres]
numpy